refactor: Question 도메인 responseDTO 생성
변경된 점:
- responseDTO 생성
